The guessed word is selected from the word dictionary and the code is changed
