+++ MindMarathon/ViewModel/FloodFillViewModel.swift
@@ -32,6 +32,8 @@ class FloodFillViewModel {
         return gameRules
     }
     
+    
+    
     func selectMaxLenght(maxLenght: String) -> String {
         var newLenght: Int = Int(maxLenght)!
